ChebyshevExpansions

Documentation for ChebyshevExpansions.

ChebyshevExpansions.ChebyshevExpansions

ChebyshevExpansions.averageKPM

ChebyshevExpansions.densityKPM

ChebyshevExpansions.dosKPM

ChebyshevExpansions.momentaKPM

ChebyshevExpansions.randomkets

ChebyshevExpansions.ChebyshevExpansions — Module

ChebyshevExpansions

ChebyExpansions.jl

Introduction

In this package, we implement the Kernel Polynomial Method (KPM) in Julia, an efficient technique to compute spectral and correlation functions using order-N Chebyshev expansions of said quantities.

Diagonalization in KPM is replaced by matrix multiplications and, thus, the resource consumption for a given precission drops to $\mathcal{O}(D)$ for sparse or $\mathcal{O}(D^2)$ for dense matrices, being $D$ the matrix dimension.

In this regard, this technique is well suited for large matrix calculations where exact diagonalization becomes too expensive. In addition, KPM offers some advantages in terms of stability and resource consumption with respect to other recursive methods such as Lanczos-based techniques, see [1] for a comprehensive review.

Practical considerations

The energy resolution is controlled by the ratio $W/N$, where $W$ is the bandwidth of $H$ and $N$ the expansion order. The complexity for sparse matrices goes like $\mathcal O(ND)$

The error in the stochastic trace estimation $\mathcal{O}(1/\sqrt{DR})$, being R the number of random kets considered.

Observables

Currently, given a hamiltonian matrix, $H$ and a generic operator $A$ the following methods are implemented:

Momenta computation corresponding to traces over the product of Chebyshev polynomials of $H$ and $A$: $\mu_n = \text{Tr}[A T_n(H)]$.

Calculation of spectral densities of an operator $A(E) = \text{Tr} [A 𝛿(E-H)]$.

Calculation of thermal averages, i.e. $\langle A\rangle _{E_F} = \text{Tr} [A f(H,E_F)]$

Exported functions

momentaKPM dosKPM computes the density of states $(A == I)$ densityKPM averageKPM

References

[1] A. Weiße, G. Wellein, A. Alvermann, and H. Fehske, The kernel polynomial method, Reviews of Modern Physics 78, 275 (2006).

ChebyshevExpansions.averageKPM — Method

`averageKPM(h::Hamiltonian, A; kBT = 0, Ef = 0, kw...)`

Compute, using the Kernel Polynomial Method (KPM), the thermal expectation value ⟨A⟩ = Σ_k f(E_k) ⟨k|A|k⟩ = ∫dE f(E) Tr [A δ(E-H)]/N₀ = Tr [A f(H)]/N₀ for a given hermitian operator A and a zero-dimensional hamiltonian h with a total of N₀ orbitals (see momentaKPM and its options kw for further details). f(E) is the Fermi-Dirac distribution function, |k⟩ are h eigenstates with energy E_k, kBTis the temperature in energy units andEf` the Fermi energy.

`averageKPM(μ::MomentaKPM, A; kBT = 0, Ef = 0)``

Same as above with the KPM momenta as input (see momentaKPM).

See also

`dosKPM`, `momentaKPM`, `averageKPM`

ChebyshevExpansions.densityKPM — Method

densityKPM(h::Hamiltonian, A; resolution = 2, kets = randomkets(1), kw...)

Compute, using the Kernel Polynomial Method (KPM), the local density of states ρₖ(ϵ) = ⟨k|δ(ϵ-h)|k⟩ for a single ket |k⟩, a collection of {|k_i⟩}, or a set of randomly- generated kets for a stochastic trace calculation(see momentaKPM).

The result is a tuple of energy points ϵᵢ::Vector spanning the band range, and real ρᴬₖ(ϵᵢ)::Vector values. The number of energy points ϵᵢ is order *resolution, rounded to the closest integer.

If kets contains more than a single ket, the sum ∑ₖρₖ(ε) over all models will be computed. In the case of the default kets = randomkets(h, I, n), this results in an estimate of the total density of states per orbital, computed through an stochastic trace, ρ(ϵ) = ∑ₖ⟨k|δ(ϵ-h)|k⟩/n ≈ Tr[δ(ϵ-h)]/N₀, where N₀ is the total number of orbitals in the unit cell.

`densityKPM(μ::MomentaKPM; resolution = 2)`

Same as above with KPM momenta μ as input.

See also

`dosKPM`, `momentaKPM`, `averageKPM`

ChebyshevExpansions.dosKPM — Method

dosKPM(h::Hamiltonian; resolution = 2, kets = randomkets(1), kw...)

Compute, using the Kernel Polynomial Method (KPM), the local density of states ρₖ(ϵ) = ⟨k|δ(ϵ-h)|k⟩ for a single ket |k⟩, a collection of {|k_i⟩}, or a set of randomly- generated kets for a stochastic trace calculation(see momentaKPM).

The result is a tuple of energy points ϵᵢ::Vector spanning the band range, and real ρₖ(ϵᵢ)::Vector values (any residual imaginary part in ρₖ is dropped). The number of energy points ϵᵢ is order * resolution, rounded to the closest integer.

If kets contains more than a single ket, the sum ∑ₖρₖ(ε) over all models will be computed. In the case of the default kets = randomkets(h, I, n), this results in an estimate of the total density of states per orbital, computed through an stochastic trace, ρ(ϵ) = ∑ₖ⟨k|δ(ϵ-h)|k⟩/n ≈ Tr[δ(ϵ-h)]/N₀, where N₀ is the total number of orbitals in the unit cell.

dosKPM is a particular case of densityKPM for an operator A = I and with any residual imaginary parts dropped

`dosKPM(μ::MomentaKPM; resolution = 2)`

Same as above with KPM momenta μ as input.

See also

`momentaKPM`, `densityKPM`, `averageKPM`

ChebyshevExpansions.momentaKPM — Function

`momentaKPM(h, A = I; kets = randomkets(h, A, n), order = 10, bandrange = missing)`

Compute the Kernel Polynomial Method (KPM) momenta μₙ = ⟨k|Tₙ(h) A|k⟩, where |k⟩ may either refer to a single ket (kets::AbstractVector), or a set of kets::AbstractMatrix stored as columns, or to a collection of n random kets generated by randomkets(h, A, n). A is an observable (AbstractMatrix) and Tₙ(h) is the order-n Chebyshev polynomial of the Hamiltonian h.

For the randomkets case, μₙ is summed over all kets, μₙ= ∑_k ⟨k|Tₙ(h) A|k⟩. If kets are a multi-column ket, a sum over columns will similarly be performed.

kets generated with randomkets() are useful to estimate momenta of normalized traces using the stochastic trace approach, whereby μ_n = Tr[A T_n(h)]/N₀ ≈ ∑ₖ⟨k|A T_n(h)|k⟩. Here the |k⟩s are n random kets of norm 1/√n and N₀ is the total number of orbitals per unit cell of h (see randomkets).

The order of the Chebyshev expansion is order. The bandbrange = (ϵmin, ϵmax) should completely encompass the full bandwidth of hamiltonian. If missing it is computed automatically using ArnoldiMethod (must be loaded using ArnoldiMethod).

Examples

julia> momentaKPM(h, I, kets = randomkets(h, I, 5), order = 20)

julia> momentaKPM(h, bandrange = (-6,6)).mulist |> length

ChebyshevExpansions.randomkets — Method

randomkets(h, A, n)

Creates n random vectors with length D = dim(h) and stacks them into a D×n matrix. The resulting ket amplitudes are normalized by 1/√n.

This type of ket structure is useful e.g. in stochastic trace evaluation of KPM methods, where the amplitude, a = cis(2pi*rand()) is chosen as a random function s.t. ⟨a⟩ = 0, ⟨aa⟩ = 0 and ⟨a'a⟩ = 1. Then, the normalized trace of an operator A can be estimated with Tr[A]/N₀ ≈ ∑⟨ket|A|ket⟩, where the sum is taken over the n random kets |ket⟩ of norm 1/√n produced by randomkets, and N₀ is the total number of orbitals in the full unit cell.

By default when applied to a multiorbital system with a maximum of N orbitals per site, the generated kets have independent, complex, normally-distributed random components